# docs/public/openapi.yaml
openapi: 3.1.0
info:
  title: Recon control plane
  version: "0.1.0"
  summary: The HTTP surface of the Recon attack surface management platform.
  description: |
    Two kinds of caller reach this API and they are told apart by the credential
    they carry, never by a field in a body.

    A **console token** is a long lived organization credential. It carries a set
    of actions (`read_assets`, `manage_scope`, `manage_jobs`) and every route it
    reaches goes through one authorization layer that turns the token into a
    principal. The organization is discovered from the token and can never be
    named by the caller: no filter, no body field and no query parameter selects
    a tenant.

    A **run token** is short lived, signed, and belongs to one run. It reaches
    exactly two routes: the target list that run is allowed to read, and the
    report it posts back. Nothing else.

    Two conventions hold across the whole surface.

    An identifier that does not exist and one belonging to another organization
    answer the same `404`. The difference between them would leak one bit, and
    one bit is enough to enumerate.

    Every error body carries a machine readable `error` beside a sentence in
    `detail`. Branch on the first, show the second.
servers:
  - url: http://localhost:8080
    description: Local development.
  - url: https://{host}
    description: A deployment.
    variables:
      host:
        default: recon.example.com
...
security:
  - consoleToken: []
paths:
  /assets/search:
    post:
      tags: [assets]
      operationId: searchAssets
      summary: One page of assets.
      description: |
        A structured filter tree rather than a query string, and a POST rather
        than a GET, because the question is a document.

        An unknown field or an operator a field does not accept is refused by
        name rather than answered with an empty result set: an empty page reads
        as an empty inventory, which is a different thing.

        Requires `read_assets`.
      security:
        - consoleToken: [read_assets]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Query'
            examples:
              everything:
                summary: The whole inventory, first page.
                value: {}
              live_http_in_one_program:
                summary: Live HTTP services of one perimeter.
                value:
                  filter:
                    op: and
                    clauses:
                      - op: eq
                        field: program_id
                        value: 6f1a2b3c-0d4e-4f5a-8b6c-7d8e9f0a1b2c
                      - op: eq
                        field: lifecycle
                        value: active
                      - op: in
                        field: status_code
                        value: [200, 401, 403]
                  limit: 100
      responses:
        '200':
          description: One page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '400':
          $ref: '#/components/responses/BadFilter'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/Unavailable'
  /assets/hosts:
    post:
      tags: [assets]
      operationId: listHosts
      summary: One page of the folded list.
      description: |
        The list a console renders: assets grouped under the host they share.

        A route of its own rather than a flag on `/assets/search`, because the
        two hand out cursors that mean different columns. Behind one route, a
        client that kept a cursor and flipped a flag would get a walk that
        restarts or skips, and neither says anything.

        Requires `read_assets`.
      security:
        - consoleToken: [read_assets]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Query'
      responses:
        '200':
          description: One page of groups.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupedPage'
        '400':
          $ref: '#/components/responses/BadFilter'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/Unavailable'
  /assets/facets:
    post:
      tags: [assets]
      operationId: facetAssets
      summary: Counters over the filtered result.
      description: |
        Aggregates over what the filter selected, not over the inventory. That
        is what the side counters of an ASM interface are.

        `limit` and `cursor` are ignored here. Requires `read_assets`.
      security:
        - consoleToken: [read_assets]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Query'
      responses:
        '200':
          description: The counters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacetPage'
        '400':
          $ref: '#/components/responses/BadFilter'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/Unavailable'
  /assets/export:
    post:
      tags: [assets]
      operationId: exportAssets
      summary: The filtered result in full, streamed.
      description: |
        It streams as it walks, so the response starts before the walk
        finishes. One consequence is worth stating plainly: a failure part way
        through has already sent a valid prefix and the status line is long
        gone, so a truncated file is possible and is logged rather than
        reported in the status.

        `jsonl` is lossless. `csv` flattens, and opens in a spreadsheet.

        `limit` here is a total rather than a page, and it is honoured exactly.
        Requires `read_assets`.
      security:
        - consoleToken: [read_assets]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/Query'
              properties:
                format:
                  type: string
                  enum: [jsonl, csv]
                  default: jsonl
      responses:
        '200':
          description: The walk, one asset per line.
          content:
            application/x-ndjson:
              schema:
                type: string
                description: One `Asset` object per line.
            text/csv:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadFilter'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/Unavailable'
  /assets/fields:
    get:
      tags: [assets]
      operationId: getFields
      summary: What the search accepts, and what this deployment can do.
      description: |
        Served rather than deduced. A client that learns the vocabulary against
        400s learns it wrong, and one that deduces the enrichment state from
        missing data guesses: with no MaxMind database configured, no asset
        carries an ASN, and that looks exactly like enrichment finding nothing.

        Call this first. It is the authoritative field list, and the static one
        in this document is a copy that can fall behind a deployment.

        Requires `read_assets`.
      security:
        - consoleToken: [read_assets]
      responses:
        '200':
          description: The vocabulary.
          content:
            application/json:
              schema:
                type: object
                properties:
                  fields:
                    type: object
                    description: Field name to the operators it accepts.
                    additionalProperties:
                      type: array
                      items:
                        type: string
                  facets:
                    type: integer
                    description: How many values one facet returns.
                  page_limit:
                    type: integer
                    description: The largest page a caller may ask for.
                  page_default:
                    type: integer
                  enrichment:
                    type: object
                    properties:
                      configured:
                        type: boolean
                        description: |
                          Whether this deployment has an enrichment database at
                          all. False means the infrastructure fields are never
                          populated, which is not the same as never matching.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /assets/{asset}:
    get:
      tags: [assets]
      operationId: getAsset
      summary: One asset, with its evidence and its timeline.
      description: |
        The only read path here that touches the journal.

        The timeline is bounded by a window and a cap, and a layer whose history
        was cut is named in `truncated_layers` rather than silently shortened.

        Requires `read_assets`.
      security:
        - consoleToken: [read_assets]
      parameters:
        - $ref: '#/components/parameters/AssetID'
      responses:
        '200':
          description: The asset.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Unavailable'
  /assets/{asset}/render:
    post:
      tags: [renders]
      operationId: requestRender
      summary: Put one asset at the head of the render queue.
      description: |
        `manage_jobs` rather than `ingest`, and the distinction is not
        tidiness: something holding `ingest` could otherwise schedule renders of
        its choosing and spend a programme's budget on targets it picked.

        A `200` does not mean a render is coming. `queued` is read from the
        conditions the statement itself applies, so an archived asset, one out
        of scope, or one on a port a browser refuses to open all answer `200`
        with `queued: false` and a reason. Answering otherwise would tell a
        caller to wait for a render nobody will make.
      security:
        - consoleToken: [manage_jobs]
      parameters:
        - $ref: '#/components/parameters/AssetID'
      responses:
        '200':
          description: Whether anything is queued.
          content:
            application/json:
              schema:
                type: object
                required: [asset_id, key, queued]
                properties:
                  asset_id:
                    type: string
                    format: uuid
                  key:
                    type: string
                  queued:
                    type: boolean
                  reason:
                    type: string
                    description: Present only when nothing was queued.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Unavailable'
  /renders/replan:
    post:
      tags: [renders]
      operationId: replanRenders
      summary: Put the whole inventory back in the low queue.
      description: |
        The forced refresh after a major update of the rendering service. It is
        spread over several days on purpose: doing it in an hour would restore
        baseline consistency and produce a mass alert at the same time.

        The spread is a deployment setting, not a parameter. Requires
        `manage_jobs`.
      security:
        - consoleToken: [manage_jobs]
      responses:
        '200':
          description: What moved.
          content:
            application/json:
              schema:
                type: object
                properties:
                  replanned:
                    type: integer
                  spread:
                    type: string
                    description: A Go duration, as configured.
                    examples: ["72h0m0s"]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/Unavailable'
  /feed:
    get:
      tags: [assets]
      operationId: streamFeed
      summary: The live feed of discoveries, as server sent events.
      description: |
        Polling with a cursor behind an SSE connection. Rounds run every few
        seconds and a round that found nothing emits nothing, so the cursor only
        advances on a message that named a discovery.

        The stream ends by itself after a few minutes and the client reconnects,
        which the SSE protocol does on its own. A forgotten tab holding an HTTP
        request forever is a leak visible only in the connection count.

        Events are named `discoveries` rather than left unnamed, so adding a
        second kind of message later does not start delivering it to every
        existing listener. Comment lines beginning with `:` are heartbeats and
        the protocol drops them.

        Resumption costs no server state: the message id *is* the cursor, and a
        browser hands it back as `Last-Event-ID` by itself. A first connection
        with no cursor starts at the present rather than at the beginning of the
        inventory.

        Requires `read_assets`.
      security:
        - consoleToken: [read_assets]
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: |
            Where to resume. Used only when `Last-Event-ID` is absent, which is
            the case of a first connection that already knows where it left off.
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: string
          description: The browser's own resumption, and it takes precedence.
      responses:
        '200':
          description: The event stream.
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Each event is `id:` the cursor, `event: discoveries`, and
                  `data:` one `Tick` object, given below.
              x-event-data-schema:
                $ref: '#/components/schemas/Tick'
        '400':
          description: The cursor is not one this served.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /programs/{program}/rules/{rule}:
    patch:
      tags: [programs]
      operationId: updateRule
      summary: Edit or close a rule.
      description: |
        Closing a rule is setting `valid_to`. There is no delete on this surface
        at all: a rule has a period of validity rather than an existence, and an
        asset classified by a rule since closed stays explainable.

        `version` is required, and a stale one is refused with `409` having
        written nothing. Requires `manage_scope`.
      security:
        - consoleToken: [manage_scope]
      parameters:
        - $ref: '#/components/parameters/ProgramID'
        - name: rule
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/RuleBody'
              required: [version]
      responses:
        '200':
          description: Written, and what it moved.
          content:
            application/json:
              schema:
                type: object
                properties:
                  rule:
                    $ref: '#/components/schemas/Rule'
                  effect:
                    $ref: '#/components/schemas/Effect'
        '400':
          description: No version, or a rule the perimeter cannot hold.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/StaleVersion'
        '500':
          $ref: '#/components/responses/Unavailable'
...
components:
  securitySchemes:
    consoleToken:
      type: http
      scheme: bearer
      description: |
        A long lived organization credential, sent as `Authorization: Bearer`.
        It names itself and never an organization: the tenant is discovered from
        the token, in one place, and everything after that runs scoped to it.

        The actions are `read_assets` (the inventory and the queues),
        `manage_scope` (perimeters, rules, hand entered assets) and
        `manage_jobs` (runs and renders). A credential that authenticated but
        lacks the action gets `403`, told apart from the `401` of a credential
        that is wrong, which is what makes the separation usable rather than
        mysterious.
    runToken:
      type: http
      scheme: bearer
      description: |
        A signed, short lived credential belonging to one run. It reaches
        `GET /runs/{run}/targets` and `POST /reports`, and nothing else. It
        cannot be recalled, so a run in a terminal state simply stops answering.
...